requirements.txt
python-telegram-bot
python-dotenv
requests
groq
cloudscraper
beautifulsoup4
